Own the cutting of a polyline into simple fragments in the geometry layer
The positions at which a polyline crosses itself come from a purely geometric
computation over its points, which lives in the temporal point file and is
reachable only through a temporal sequence, so nothing outside that file asks
for it. The segment intersection it rests on, and the two box helpers, have
no other caller.

pointarr_find_splits answers those positions for an array of points, and
geo_funcs.c holds it together with what it rests on.
tpointseq_linear_find_splits collects the points of the sequence and asks it,
which is the whole of what it does. The two answer the same fragments for
every temporal point of the regression suite and for eight hundred random
polylines on a grid small enough that they cross and repeat themselves often,
non-simple in two of every three cases.
